use inlined-fns approach
